apps/breeds/views.py
from django.shortcuts import render
from rest_framework import serializers, viewsets

from .models import Breed


class BreedSerializer(serializers.ModelSerializer):
    class Meta:
        model = Breed
        fields = ['id', 'name', 'origin', 'body_type', 'pattern']


class BreedViewSet(viewsets.ModelViewSet):
    queryset = Breed.objects.all()
    serializer_class = BreedSerializer
